ADD: profile controller/configured controller

+++ app/controllers/profiles_controller.rb
@@ -1,41 +1,69 @@
 class ProfilesController < ApplicationController
   before_action :authenticate_request
-  before_action :set_pet, only: [:create, :show, :update, :destroy]
+  before_action :set_owner, only: [:create, :destroy]
+  before_action :set_owner, only: [:create, :destroy]
+  before_action :set_profile, only: [:show, :update, :destroy]
 
   def show
-    profile = @pet.profile
-    render json: profile, status: :ok
+    render json: @profile, status: :ok
   end
 
+  # builds pet profile if associated pet_id is found, otherwise builds owner profile
+  # build_profile automatically handles associations
   def create
-    profile = @pet.build_profile(profile_params)
-    if profile.save
-      render json: profile, status: :created
+    if params[:pet_id].present?
+      @pet = Pet.find(params[:pet_id])
+      @profile = @pet.build_profile(profile_params)
     else
-      render json: profile.errors, status: :unprocessable_entity
+      @profile = @owner.build_profile(profile_params)
+    end
+
+    if @profile.save
+      render json: @profile, status: :created
+    else
+      render json: @profile.errors, status: :unprocessable_entity
     end
   end
 
   def update
-    if @pet.profile.update(profile_params)
-      render json: @pet.profile, status: :ok
+    if @profile.update(profile_params)
+      render json: @profile, status: :ok
     else
-      render json: @pet.profile.errors, status: :unprocessable_entity
+      render json: @profile.errors, status: :unprocessable_entity
     end
   end
 
+  # destroy profile associated with pet if pet_id provided, otherwise destroys owner profile
   def destroy
-    if @pet.profile.destroy
-      render json: nil, status: :ok
+    if params[:pet_id].present?
+      @pet = Pet.find(params[:pet_id])
+      if @pet.profile.destroy
+        render json: nil, status: :ok
     else
-      render json: @pet.profile.errors, status: :unprocessable_entity
+        render json: @pet.profile.errors, status: :unprocessable_entity
+    end
+  else
+    if @owner.profile.destroy
+      render json: nil, status: :ok
+      else
+        render json: @owner.profile.errors, status: :unprocessable_entity
+      end
     end
   end
 
   private 
-  
+
   def set_pet
     @pet = Pet.find(params[:pet_id])
+  end
+
+  def set_owner
+    @owner = Owner.find(params[:owner_id])
+  end
+
+  # checks profile associated with pet, otherwise checks for owner profile
+  def set_profile
+    @profile = params[:pet_id] ? Profile.find_by(pet_id: params[:pet_id]) : @owner.profile
   end
 
   def profile_params
